MdeModulePkg/FileSystem/UdfDxe: Support ".." in UdfOpen()
If a ".." is given in the filename passed to UdfOpen(), it should go a
directory behind the current parent directory.

As per ECMA-167 specification, FIDs contain ICBs from their parent
directories, so it will have an ICB location of its parent directory --
which made it easier to handle with.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Paulo Alcantara <pcacjr@zytor.com>