MdeModulePkg/FileSystem/UdfDxe: Fix wrong setting of file size
When UdfRead() is called with a directory it must find the current
directory by finding FID and FE of the respective entry.

This patch fixes a bug in UdfRead() function that did not read FE of the
found current directory entry -- it used erroneously the Parent FE
instead.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Paulo Alcantara <pcacjr@zytor.com>